Add files via upload
Write a program that allows users to create a music playlist consisting of n songs and then display the list of the songs. The program should prompt user to enter the name of each song and add it to the playlist in a sequential order. The playlist should be designed to repeat once it reaches the end of the list.